zcash_history: move tests to their own file
