+++ zebra-chain/src/primitives/zcash_history/tests.rs
@@ -1,0 +1,4 @@
+//! Tests for Zebra history trees
+
+#[cfg(test)]
+mod vectors;
